+++ workspace/client/src/pages/tour/Tour.js
@@ -4,7 +4,7 @@ import MobileView from "./MobileView";
 import DesktopView from "./DesktopView";
 
 const Tour = () => {
-  const isMobile = useMediaQuery({ maxWidth: 414 });
+  const isMobile = useMediaQuery({ maxWidth: 430 });
 
   return (
     <>
